feat: added more http status and default error pages if body is not passed

+++ includes/Response.hpp
@@ -9,9 +9,11 @@
 #include <ctime>
 #include <sys/stat.h>
 #include <map>
+#include "errorPages.hpp"
 
 #define SERVER_NAME "webserver/1.0"
 #define HTTP_VERSION "HTTP/1.1"
+
 
 typedef std::pair<std::string, std::string> t_fields;
 
@@ -31,8 +33,8 @@ class Response {
 		void	_serverError();
 	public:
 		Response();
-		Response(short int status);
-		Response(short int status, std::string bodyFile);
+		Response(int status);
+		Response(int status, std::string bodyFile);
 		~Response();
 		std::string	response() const;
 		std::string	getContentType(const std::string &filename) const;
@@ -43,8 +45,7 @@ class Response {
 };
 
 template<typename T>
-std::string	toString(const T& value)
-{
+std::string	toString(const T& value) {
 	std::ostringstream oss;
 	oss << value;
 	return (oss.str());
